Moved ISO_8601_FORMAT to core app
This constant is shared across multiple apps, not just payment, so should be in a more centralized location.

XCOM-505
