+++ ecommerce/core/constants.py
@@ -1,5 +1,7 @@
 """Health check constants."""
 
+ISO_8601_FORMAT = u'%Y-%m-%dT%H:%M:%SZ'
+
 
 class Status(object):
     """Health statuses."""
