+++ ecommerce/extensions/payment/constants.py
@@ -1,7 +1,5 @@
 """Payment processor constants."""
 
-ISO_8601_FORMAT = u'%Y-%m-%dT%H:%M:%SZ'
-
 CARD_TYPES = {
     'american_express': {
         'display_name': 'American Express',
